+++ source-code/sdk-js/src/adapter-sveltekit/ast-transforms/_.server.js.ts
@@ -1,4 +1,4 @@
-import { assertNoImportsFromSdkJs } from "../../ast-transforms/assertions.js"
+import { assertNoImportsFromSdkJs } from "../../ast-transforms/utils/assertions.js"
 import { codeToSourceFile } from "../../ast-transforms/utils/js.util.js"
 import type { TransformConfig } from "../vite-plugin/config.js"
 import { filePathForOutput } from "../vite-plugin/fileInformation.js"
